Bug fix: prevent players from building units during replay
Currently, if the game is not at the current action (i.e. in replay mode), and you go to a previous move of yours, you can build units on dropzones and the action is saved, and you can seek back an action and repeat. When you refresh the page, you end up with many free units on the same dropzone, basically letting you cheat :P